requirements.txt
annotated-types==0.7.0
anyio==4.4.0
assemblyai==0.30.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
exceptiongroup==1.2.2
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.0.3
fsspec==2024.10.0
future==1.0.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.0
openai-whisper @ git+https://github.com/openai/whisper.git@271445b2f24f00f8175c4fb7ae91876f7451dfc1
packaging==24.1
pillow==10.4.0
proglog==0.1.10
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
websockets==12.0
Werkzeug==3.0.3
whisper==1.1.10
